# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MLK Pro API",
    "version": "v1",
    "description": "Auto-generated from routes/api.php (internal API v1)."
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Local dev"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "TaskController@index",
        "operationId": "TaskController_index_get",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\TaskController@index",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete",
          "App\\Http\\Middleware\\EnsureCompanyFeature:tasks"
        ]
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "TaskController@store",
        "operationId": "TaskController_store_post",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\TaskController@store",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete",
          "App\\Http\\Middleware\\EnsureCompanyFeature:tasks"
        ]
      }
    },
    "/api/v1/tasks/{task}": {
      "put": {
        "tags": [
          "tasks"
        ],
        "summary": "TaskController@update",
        "operationId": "TaskController_update_put",
        "parameters": [
          {
            "name": "task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\TaskController@update",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete",
          "App\\Http\\Middleware\\EnsureCompanyFeature:tasks"
        ]
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "TaskController@destroy",
        "operationId": "TaskController_destroy_delete",
        "parameters": [
          {
            "name": "task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\TaskController@destroy",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete",
          "App\\Http\\Middleware\\EnsureCompanyFeature:tasks"
        ]
      }
    },
...
  }
}